+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Ensure active users have a confirmed_at date (migration) [#298](https://github.com/etalab/udata-gouvfr/pull/298)
 - Remove credits page [#306](https://github.com/etalab/udata-gouvfr/pull/306)
 - Fix `modal-lg` width [#311](https://github.com/etalab/udata-gouvfr/pull/311)
+- Add tracking and privacy page [#310](https://github.com/etalab/udata-gouvfr/pull/310)
 
 ## 1.3.2 (2018-03-28)